refactor: 컴포넌트 TeamMember 리팩터링
