--- src/components/TeamListItem/index.ts
@@ -1,1 +1,0 @@
-export { default as TeamListItem } from './TeamListItem';
