+++ src/components/TeamMember/TeamMember.styles.ts
@@ -1,6 +1,7 @@
 import styled from '@emotion/styled';
 
-export const TeamListItemContainer = styled.div`
+export const Anchor = styled.div`
+  text-decoration: none;
   display: flex;
   flex-direction: row;
 `;
