+++ src/components/TeamMember/index.ts
@@ -1,0 +1,1 @@
+export { default as TeamMember } from './TeamMember';
